Change GridCoordinate from isize to i32.
I was working on using fewer sloppy arithmetic conversions, and I
realized that it's a particularly bad idea for the basic game data
types to be different on different platforms (e.g. 64-bit server,
wasm32 client.)

+++ all-is-cubes/src/math.rs
@@ -12,7 +12,7 @@ use std::hash::{Hash, Hasher};
 use std::ops::{Add, AddAssign, Div, Index, IndexMut, Mul, Rem};
 
 /// Coordinates that are locked to the cube grid.
-pub type GridCoordinate = isize;
+pub type GridCoordinate = i32;
 /// Positions that are locked to the cube grid.
 pub type GridPoint = Point3<GridCoordinate>;
 /// Vectors that are locked to the cube grid.
